pyproject.toml
[tool.poetry]
name = "datafaker"
version = "0.1"
description = "Generates fake SQL data"
authors = ["Tim Band <3266052+tim-band@users.noreply.github.com>"]
license = "MIT"
readme = "README.md"
classifiers = [
    "Environment :: Console",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Topic :: Database",
    "Topic :: Utilities",
]
homepage = "https://github.com/SAFEHR-data/datafaker"

[tool.poetry.dependencies]
python = ">=3.10,<3.14"
pydantic = {extras = ["dotenv"], version = "^1.10.2"}
psycopg2-binary = "^2.9.5"
sqlalchemy-utils = "^0.41.2"
mimesis = "^18.0.0"
typer = "^0.15.4"
pyyaml = "^6.0"
sqlalchemy = "^2"
sphinx-rtd-theme = {version = "^1.2.0", optional = true}
sphinxcontrib-napoleon = {version = "^0.7", optional = true}
smartnoise-sql = "^1"
jinja2 = "^3.1.2"
black = "^23.3.0"
jsonschema = "^4.17.3"
asyncpg = "^0.30.0"
greenlet = "^3.1.1"
pymysql = "^1.1.0"
pandas = "^2"
parsy = "^2.1"
prettytable = "^3.15.1"

[tool.poetry.group.dev.dependencies]
isort = "^5.10.1"
pylint = "^3"
mypy = "^1.5"
types-pyyaml = "^6.0.12.4"
pydocstyle = "^6.3.0"
restructuredtext-lint = "^1.4.0"
pygments = "^2.14.0"
rstcheck-core = {extras = ["sphinx"], version = "^1.0.3"}
json-schema-for-humans = "^1.3.4"
pre-commit = "^3.3.3"
testing-postgresql = "^1.3.0"

[tool.poetry.group.extras.dependencies]
tqdm = "^4.65.0"

[tool.poetry.extras]
docs = ["sphinx-rtd-theme", "sphinxcontrib-napoleon"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
datafaker = "datafaker.main:app"

[tool.isort]
profile = "black"
